refactor the app completely
