+++ app/src/components/Header/index.ts
@@ -1,0 +1,12 @@
+export { Header as default, Header } from "./Header";
+export type {
+  HeaderProps,
+  SearchBarProps,
+  SelectedDropdownProps,
+  SessionButtonsProps,
+} from "./types";
+export { SearchBar } from "./SearchBar";
+export { SelectedDropdown } from "./SelectedDropdown";
+export { SessionButtons } from "./SessionButtons";
+
+
