+++ app/src/components/InteractiveLegend/InteractiveLegend.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import React, {
+import {
   useState,
   useEffect,
   useRef,
@@ -8,148 +8,12 @@ import React, {
   forwardRef,
   useImperativeHandle,
 } from "react";
-import * as d3 from "d3";
-import html2canvas from "html2canvas";
-
-// Define the same SHAPE_MAPPING as in ImprovedScatterplot.tsx for consistency
-const SHAPE_MAPPING = {
-  asterisk: d3.symbolAsterisk,
-  circle: d3.symbolCircle,
-  cross: d3.symbolCross,
-  diamond: d3.symbolDiamond,
-  plus: d3.symbolPlus,
-  square: d3.symbolSquare,
-  star: d3.symbolStar,
-  triangle: d3.symbolTriangle,
-  wye: d3.symbolWye,
-  times: d3.symbolTimes,
-} as const;
-
-// Add these constants at the top of the file after imports
-const DEFAULT_STYLES = {
-  other: {
-    color: "#888888",
-    shape: "circle",
-  },
-  null: {
-    color: "#888888",
-    shape: "circle",
-  },
-};
-
-export interface LegendItem {
-  value: string | null;
-  color: string;
-  shape: string;
-  count: number;
-  isVisible: boolean;
-  // Add z-order for controlling the layering of items
-  zOrder: number;
-  // Flag for items that were extracted from "Other"
-  extractedFromOther?: boolean;
-}
-
-export interface InteractiveLegendProps {
-  featureData: {
-    name: string;
-    values: (string | null)[];
-    colors: string[];
-    shapes: string[];
-  };
-  featureValues: (string | null)[];
-  proteinIds?: string[];
-  maxVisibleValues?: number;
-  onToggleVisibility?: (value: string | null) => void;
-  onExtractFromOther?: (value: string) => void;
-  onSetZOrder?: (zOrderMapping: Record<string, number>) => void;
-  onOpenCustomization?: () => void;
-  selectedItems?: string[];
-  className?: string;
-  isolationMode?: boolean;
-  splitHistory?: string[][];
-}
-
-interface OtherItemsDialogProps {
-  otherItems: [string | null, number][];
-  onExtractItem: (value: string) => void;
-  onClose: () => void;
-}
-
-// Component for dialog to extract values from "Other" category
-const OtherItemsDialog: React.FC<OtherItemsDialogProps> = ({
-  otherItems,
-  onExtractItem,
-  onClose,
-}) => {
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black bg-opacity-50">
-      <div className="bg-white dark:bg-gray-800 p-6 rounded-lg shadow-xl max-w-md w-full max-h-[80vh] overflow-auto">
-        <div className="flex justify-between items-center mb-4">
-          <h3 className="text-lg font-medium">
-            Extract from &apos;Other&apos; category
-          </h3>
-          <button
-            onClick={onClose}
-            className="text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-200"
-          >
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              className="h-6 w-6"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M6 18L18 6M6 6l12 12"
-              />
-            </svg>
-          </button>
-        </div>
-
-        <div className="mb-4">
-          <p className="text-sm text-gray-600 dark:text-gray-300">
-            Select items to extract from the &apos;Other&apos; category.
-            Extracted items will appear individually in the legend.
-          </p>
-        </div>
-
-        <ul className="divide-y divide-gray-200 dark:divide-gray-700 mb-4 max-h-60 overflow-y-auto">
-          {otherItems.map(([value, count]) => (
-            <li
-              key={value === null ? "null" : value}
-              className="py-2 flex justify-between items-center hover:bg-gray-50 dark:hover:bg-gray-700 px-2 rounded"
-            >
-              <div className="flex items-center">
-                <span>{value === null ? "N/A" : value}</span>
-                <span className="ml-2 text-xs text-gray-500 dark:text-gray-400">
-                  ({count})
-                </span>
-              </div>
-              <button
-                onClick={() => value !== null && onExtractItem(value)}
-                className="text-blue-500 hover:text-blue-600 text-sm font-medium"
-              >
-                Extract
-              </button>
-            </li>
-          ))}
-        </ul>
-
-        <div className="flex justify-end">
-          <button
-            onClick={onClose}
-            className="bg-gray-200 hover:bg-gray-300 dark:bg-gray-700 dark:hover:bg-gray-600 text-gray-800 dark:text-gray-200 px-4 py-2 rounded"
-          >
-            Close
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-};
+import OtherItemsDialog from "./OtherItemsDialog";
+import { buildFrequencyMap, computeLegendItems, toZOrderMap } from "./utils";
+import { InteractiveLegendProps, LegendItem } from "./types";
+import { useLegendExport } from "./hooks/useLegendExport";
+import { useLegendDnD } from "./hooks/useLegendDnD";
+import LegendListItem from "./LegendListItem";
 
 const InteractiveLegend = forwardRef<
   { downloadAsImage: () => Promise<void> },
@@ -176,180 +40,68 @@ const InteractiveLegend = forwardRef<
     const [otherItems, setOtherItems] = useState<[string | null, number][]>([]);
     const [showOtherDialog, setShowOtherDialog] = useState(false);
 
-    // Drag state
-    const [draggedItem, setDraggedItem] = useState<string | null>(null);
-    const dragTimeoutRef = useRef<NodeJS.Timeout | null>(null);
+    // Export hook
+    const { componentRef, downloadAsImage } = useLegendExport();
+
+    // Drag-and-drop hook
+    const { draggedItem, handleDragStart, handleDragOver, handleDragEnd } = useLegendDnD((updated) => {
+      if (onSetZOrder) onSetZOrder(toZOrderMap(updated));
+      setLegendItems(updated);
+    });
 
     // Process data into legend items
     useEffect(() => {
       if (!featureData || !featureValues || featureValues.length === 0) {
+        setOtherItems([]);
         setLegendItems([]);
         return;
       }
 
-      // Create a map of value frequencies
-      const frequencyMap = new Map<string | null, number>();
-
-      // Filter featureValues based on split history when in isolation mode
-      const filteredIndices = new Set<number>();
-
-      if (
-        isolationMode &&
-        splitHistory &&
-        splitHistory.length > 0 &&
+      const frequencyMap = buildFrequencyMap(
+        featureValues,
+        isolationMode,
+        splitHistory,
         proteinIds
-      ) {
-        // First, identify which indices to include based on the split history
-        proteinIds.forEach((id, index) => {
-          // For the first split, check if the protein is in the first selection
-          let isIncluded = splitHistory[0].includes(id);
+      );
 
-          // For each subsequent split, check if the protein is also in that selection
-          if (isIncluded && splitHistory.length > 1) {
-            for (let i = 1; i < splitHistory.length; i++) {
-              if (!splitHistory[i].includes(id)) {
-                isIncluded = false;
-                break;
-              }
+      setLegendItems((previousItems) => {
+        const { items, otherItems } = computeLegendItems({
+          featureData: {
+            values: featureData.values,
+            colors: featureData.colors,
+            shapes: featureData.shapes,
+          },
+          frequencyMap,
+          maxVisibleValues,
+          includeOther: !isolationMode,
+          previousItems,
+        });
+
+        setOtherItems(otherItems);
+
+        // Shallow equality check to avoid unnecessary state churn
+        if (items.length === previousItems.length) {
+          let allEqual = true;
+          for (let i = 0; i < items.length; i++) {
+            const a = items[i];
+            const b = previousItems[i];
+            if (
+              a.value !== b.value ||
+              a.isVisible !== b.isVisible ||
+              a.zOrder !== b.zOrder ||
+              a.color !== b.color ||
+              a.shape !== b.shape ||
+              a.count !== b.count
+            ) {
+              allEqual = false;
+              break;
             }
           }
-
-          if (isIncluded) {
-            filteredIndices.add(index);
-          }
-        });
-      }
-
-      // Count frequencies of the filtered values
-      if (isolationMode && splitHistory && splitHistory.length > 0) {
-        // Only count values from proteins that pass the split filter
-        featureValues.forEach((value, index) => {
-          if (filteredIndices.has(index)) {
-            frequencyMap.set(value, (frequencyMap.get(value) || 0) + 1);
-          }
-        });
-      } else {
-        // Count all values when not in isolation mode
-        featureValues.forEach((value) => {
-          frequencyMap.set(value, (frequencyMap.get(value) || 0) + 1);
-        });
-      }
-
-      // Convert to array and sort by frequency (descending)
-      const sortedItems = Array.from(frequencyMap.entries()).sort(
-        (a, b) => b[1] - a[1]
-      ); // Sort by count, descending
-
-      // When in isolation mode, we only show the values that actually appear in the data
-      // This makes the legend more relevant to what's currently displayed
-      const filteredSortedItems = isolationMode
-        ? sortedItems.filter(([value]) => frequencyMap.has(value))
-        : sortedItems;
-
-      // Take the top N items
-      const topItems = filteredSortedItems.slice(0, maxVisibleValues);
-
-      // Find null entry
-      const nullEntry = filteredSortedItems.find(([value]) => value === null);
-
-      // Get items that will go into the "Other" category (excluding null)
-      const otherItemsArray = filteredSortedItems
-        .slice(maxVisibleValues)
-        .filter(([value]) => value !== null);
-
-      // Store "Other" items for the dialog
-      setOtherItems(otherItemsArray);
-
-      // Calculate count for "Other" category
-      const otherCount = otherItemsArray.reduce(
-        (sum, [, count]) => sum + count,
-        0
-      );
-
-      // Create legend items with z-order
-      const items: LegendItem[] = topItems.map(([value, count], index) => {
-        const valueIndex =
-          value !== null
-            ? featureData.values.indexOf(value)
-            : featureData.values.findIndex((v) => v === null);
-
-        return {
-          value,
-          color:
-            valueIndex !== -1
-              ? featureData.colors[valueIndex]
-              : DEFAULT_STYLES.null.color,
-          shape:
-            valueIndex !== -1
-              ? featureData.shapes[valueIndex]
-              : DEFAULT_STYLES.null.shape,
-          count,
-          isVisible: true,
-          zOrder: index,
-        };
-      });
-
-      // Add "Other" if needed and if we're not in isolation mode
-      // In isolation mode, we generally want to show all values explicitly
-      if (otherCount > 0 && !isolationMode) {
-        items.push({
-          value: "Other",
-          color: DEFAULT_STYLES.other.color,
-          shape: DEFAULT_STYLES.other.shape,
-          count: otherCount,
-          isVisible: true,
-          zOrder: items.length,
-        });
-      }
-
-      // Add null if not already included in top items
-      if (nullEntry && !topItems.some(([value]) => value === null)) {
-        const valueIndex = featureData.values.findIndex((v) => v === null);
-        items.push({
-          value: null,
-          color:
-            valueIndex !== -1
-              ? featureData.colors[valueIndex]
-              : DEFAULT_STYLES.null.color,
-          shape:
-            valueIndex !== -1
-              ? featureData.shapes[valueIndex]
-              : DEFAULT_STYLES.null.shape,
-          count: nullEntry[1],
-          isVisible: true,
-          zOrder: items.length,
-        });
-      }
-
-      // Get previously extracted items
-      const extractedItems = legendItems.filter(
-        (item) => item.extractedFromOther
-      );
-
-      // Add extracted items, but only if they exist in the current data (important for isolation mode)
-      extractedItems.forEach((extractedItem) => {
-        // Only add if not already in the list and if they exist in the current frequencies
-        if (
-          !items.some((item) => item.value === extractedItem.value) &&
-          frequencyMap.has(extractedItem.value)
-        ) {
-          // Find the original frequency of this item
-          const itemFrequency = filteredSortedItems.find(
-            ([value]) => value === extractedItem.value
-          );
-
-          if (itemFrequency) {
-            items.push({
-              ...extractedItem,
-              count: itemFrequency[1],
-              zOrder: items.length,
-            });
-          }
+          if (allEqual) return previousItems;
         }
-      });
 
-      // Set items state
-      setLegendItems(items);
+        return items;
+      });
     }, [
       featureData,
       featureValues,
@@ -465,262 +217,15 @@ const InteractiveLegend = forwardRef<
       [otherItems, featureData, legendItems.length, onExtractFromOther]
     );
 
-    // Simple drag and drop implementation
-    const handleDragStart = useCallback((item: LegendItem) => {
-      setDraggedItem(item.value);
-
-      // Clear any existing timeout
-      if (dragTimeoutRef.current) {
-        clearTimeout(dragTimeoutRef.current);
-      }
-    }, []);
-
-    // Handle element drag over
-    const handleDragOver = useCallback(
-      (item: LegendItem) => {
-        if (!draggedItem || draggedItem === item.value) return;
-
-        // Use a debounced approach to prevent too many re-renders
-        if (dragTimeoutRef.current) {
-          clearTimeout(dragTimeoutRef.current);
-        }
-
-        dragTimeoutRef.current = setTimeout(() => {
-          setLegendItems((prev) => {
-            // Find the indices
-            const draggedIdx = prev.findIndex((i) => i.value === draggedItem);
-            const targetIdx = prev.findIndex((i) => i.value === item.value);
-            if (draggedIdx === -1 || targetIdx === -1) return prev;
-
-            // Create a new array with the item moved
-            const newItems = [...prev];
-            const [movedItem] = newItems.splice(draggedIdx, 1);
-            newItems.splice(targetIdx, 0, movedItem);
-
-            // Update z-order
-            const updatedItems = newItems.map((item, idx) => ({
-              ...item,
-              zOrder: idx,
-            }));
-
-            // Notify parent of z-order change
-            if (onSetZOrder) {
-              const zOrderMap = updatedItems.reduce((acc, item) => {
-                if (item.value !== null) {
-                  acc[item.value] = item.zOrder;
-                }
-                return acc;
-              }, {} as Record<string, number>);
-
-              onSetZOrder(zOrderMap);
-            }
-
-            return updatedItems;
-          });
-        }, 100);
-      },
-      [draggedItem, onSetZOrder]
+    // Adapt DnD hook to local state setter signature
+    const onDragOverItem = useCallback(
+      (item: LegendItem) => handleDragOver(item, setLegendItems),
+      [handleDragOver]
     );
 
-    const handleDragEnd = useCallback(() => {
-      setDraggedItem(null);
-
-      // Clear timeout if any
-      if (dragTimeoutRef.current) {
-        clearTimeout(dragTimeoutRef.current);
-        dragTimeoutRef.current = null;
-      }
-    }, []);
-
-    // Symbol rendering function using D3 symbols for consistency with scatterplot
-    const renderSymbol = (
-      shape: string | null,
-      color: string,
-      size = 16,
-      isSelected = false
-    ) => {
-      const halfSize = size / 2;
-
-      // Safely handle null or undefined shape
-      const shapeKey = (
-        shape || "circle"
-      ).toLowerCase() as keyof typeof SHAPE_MAPPING;
-
-      // Get the D3 symbol type (default to circle if not found)
-      const symbolType = SHAPE_MAPPING[shapeKey] || d3.symbolCircle;
-
-      // Generate the SVG path using D3
-      const path = d3
-        .symbol()
-        .type(symbolType)
-        .size(size * 8)(); // Size multiplier to make it fit well in the legend
-
-      // Some symbol types should be rendered as outlines only
-      const isOutlineOnly =
-        shapeKey === "plus" ||
-        shapeKey === "asterisk" ||
-        String(shapeKey).includes("_stroke");
-
-      // Determine stroke width based on selection state
-      const strokeWidth = isSelected ? 2 : 1;
-
-      // Determine stroke color based on selection state
-      const strokeColor = isSelected ? "#3B82F6" : "#333";
-
-      // Render the symbol
-      return (
-        <svg width={size} height={size} className="inline-block">
-          <g transform={`translate(${halfSize}, ${halfSize})`}>
-            <path
-              d={path || ""}
-              fill={isOutlineOnly ? "none" : color}
-              stroke={isOutlineOnly ? color : strokeColor}
-              strokeWidth={isOutlineOnly ? 2 : strokeWidth}
-            />
-          </g>
-        </svg>
-      );
-    };
-
-    // Add reference to the component's DOM element
-    const componentRef = useRef<HTMLDivElement>(null);
-
-    // Expose downloadAsImage method to parent component
+    // Expose downloadAsImage method to parent
     useImperativeHandle(ref, () => ({
-      downloadAsImage: async () => {
-        if (!componentRef.current) return;
-
-        try {
-          // Clone the component for processing
-          const clone = componentRef.current.cloneNode(true) as HTMLElement;
-
-          // Process all elements to handle modern color formats
-          const processColors = (element: HTMLElement) => {
-            // Apply computed styles to all children recursively
-            Array.from(element.querySelectorAll("*")).forEach((el) => {
-              if (!(el instanceof HTMLElement)) return;
-
-              try {
-                // Get computed colors
-                const computedStyle = window.getComputedStyle(el);
-
-                // Apply computed colors directly to override any modern formats
-                const colorProps = [
-                  "color",
-                  "background-color",
-                  "border-color",
-                  "fill",
-                  "stroke",
-                ];
-                colorProps.forEach((prop) => {
-                  const value = computedStyle.getPropertyValue(prop);
-                  if (value && value !== "none" && value !== "transparent") {
-                    el.style.setProperty(prop, value, "important");
-                  }
-                });
-
-                // Remove any oklch from inline styles
-                const style = el.getAttribute("style");
-                if (
-                  style &&
-                  (style.includes("oklch") ||
-                    style.includes("lab(") ||
-                    style.includes("lch(") ||
-                    style.includes("color("))
-                ) {
-                  let newStyle = style;
-                  [
-                    /oklch\([^)]+\)/g,
-                    /lab\([^)]+\)/g,
-                    /lch\([^)]+\)/g,
-                    /color\([^)]+\)/g,
-                  ].forEach((regex) => {
-                    newStyle = newStyle.replace(regex, "rgb(0,0,0)");
-                  });
-
-                  if (newStyle !== style) {
-                    el.setAttribute("style", newStyle);
-                  }
-                }
-              } catch {
-                // Silent error - continue processing
-              }
-            });
-
-            return element;
-          };
-
-          // Process the clone
-          const processedElement = processColors(clone);
-
-          // Create a container for the processed element
-          const container = document.createElement("div");
-          container.style.position = "absolute";
-          container.style.left = "-9999px";
-          container.style.top = "-9999px";
-          container.style.width = `${componentRef.current.clientWidth}px`;
-          container.style.height = `${componentRef.current.clientHeight}px`;
-          container.appendChild(processedElement);
-          document.body.appendChild(container);
-
-          // Use html2canvas to render to PNG
-          const canvas = await html2canvas(processedElement, {
-            backgroundColor: "#ffffff",
-            scale: 2, // Higher resolution
-            logging: false,
-            useCORS: true,
-            allowTaint: true,
-            onclone: (_, element) => {
-              // Final processing in cloned document
-              const elements = element.querySelectorAll("*");
-              elements.forEach((el) => {
-                if (!(el instanceof HTMLElement)) return;
-
-                try {
-                  // Handle any remaining modern color formats
-                  const style = el.getAttribute("style");
-                  if (
-                    style &&
-                    (style.includes("oklch") ||
-                      style.includes("lab(") ||
-                      style.includes("lch(") ||
-                      style.includes("color("))
-                  ) {
-                    let newStyle = style;
-                    [
-                      /oklch\([^)]+\)/g,
-                      /lab\([^)]+\)/g,
-                      /lch\([^)]+\)/g,
-                      /color\([^)]+\)/g,
-                    ].forEach((regex) => {
-                      newStyle = newStyle.replace(regex, "rgb(0,0,0)");
-                    });
-
-                    if (newStyle !== style) {
-                      el.setAttribute("style", newStyle);
-                    }
-                  }
-                } catch {
-                  // Silent error - continue processing
-                }
-              });
-              return element;
-            },
-          });
-
-          // Convert to PNG and download
-          const dataUrl = canvas.toDataURL("image/png", 1.0);
-          const link = document.createElement("a");
-          link.href = dataUrl;
-          link.download = "legend.png";
-          link.click();
-
-          // Clean up
-          document.body.removeChild(container);
-        } catch (error) {
-          console.error("Error exporting legend:", error);
-        }
-      },
+      downloadAsImage,
     }));
 
     // Sort items by z-order
@@ -728,19 +233,16 @@ const InteractiveLegend = forwardRef<
       (a, b) => a.zOrder - b.zOrder
     );
 
-    // At the bottom of the file, add displayName
-    InteractiveLegend.displayName = "InteractiveLegend";
-
     return (
       <div
-        className={`interactive-legend-component p-3 border rounded-md shadow-sm bg-white dark:bg-gray-800 dark:border-gray-700 ${className}`}
+        className={`interactive-legend-component p-3 border rounded-md shadow-sm bg-white ${className}`}
         ref={componentRef}
       >
         <div className="flex justify-between items-center mb-3">
           <h3 className="font-medium">{featureData.name}</h3>
           <button
             onClick={onOpenCustomization}
-            className="text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-200"
+            className="text-gray-500 hover:text-gray-700"
             title="Customize Legend"
           >
             <svg
@@ -782,7 +284,7 @@ const InteractiveLegend = forwardRef<
                 selectedItems.includes(item.value));
 
             return (
-              <li
+              <LegendListItem
                 key={
                   item.value === null
                     ? "null"
@@ -790,88 +292,16 @@ const InteractiveLegend = forwardRef<
                     ? "other"
                     : item.value
                 }
-                className={`
-            flex items-center justify-between p-2.5 rounded-lg cursor-pointer transition-all duration-200
-            ${
-              item.isVisible
-                ? "bg-gray-50 dark:bg-gray-800"
-                : "bg-gray-100 dark:bg-gray-700 opacity-50"
-            }
-            ${
-              draggedItem === item.value ? "bg-blue-50 dark:bg-blue-900/30" : ""
-            }
-
-            ${isItemSelected ? "ring-2 ring-red-500 dark:ring-red-400" : ""}
-            ${
-              item.extractedFromOther
-                ? "border-l-4 border-green-500 dark:border-green-400"
-                : ""
-            }
-            hover:bg-gray-100 dark:hover:bg-gray-700
-            active:bg-blue-100 dark:active:bg-blue-900/50
-            shadow-sm hover:shadow-md
-          `}
-                onClick={() => handleItemClick(item.value)}
-                onDoubleClick={() => handleItemDoubleClick(item.value)}
-                draggable={true}
-                onDragStart={() => handleDragStart(item)}
-                onDragOver={() => handleDragOver(item)}
+                item={item}
+                isSelected={isItemSelected}
+                draggedItem={draggedItem}
+                onClick={handleItemClick}
+                onDoubleClick={handleItemDoubleClick}
+                onDragStart={handleDragStart}
+                onDragOver={onDragOverItem}
                 onDragEnd={handleDragEnd}
-              >
-                <div className="flex items-center">
-                  <div
-                    className="mr-3 cursor-grab p-1 rounded flex items-center"
-                    onMouseDown={(e) => e.stopPropagation()}
-                  >
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      className="h-4 w-4 text-gray-400"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                      stroke="currentColor"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth={2}
-                        d="M4 8h16M4 16h16"
-                      />
-                    </svg>
-                  </div>
-                  <div className="mr-2">
-                    {item.value === "Other" || !item.isVisible
-                      ? renderSymbol("circle", item.isVisible ? "#888" : "#ccc")
-                      : renderSymbol(
-                          item.shape,
-                          item.color,
-                          16,
-                          isItemSelected
-                        )}
-                  </div>
-                  <span>
-                    {item.value === null
-                      ? "N/A"
-                      : item.value === "Other"
-                      ? item.value
-                      : item.value}
-                  </span>
-                  {item.value === "Other" && (
-                    <button
-                      className="ml-1 text-blue-500 hover:text-blue-600 text-xs font-medium"
-                      onClick={(e) => {
-                        e.stopPropagation();
-                        setShowOtherDialog(true);
-                      }}
-                      title="Extract items from Other"
-                    >
-                      (view)
-                    </button>
-                  )}
-                </div>
-                <span className="text-sm text-gray-500 dark:text-gray-400">
-                  {item.count}
-                </span>
-              </li>
+                onOpenOther={() => setShowOtherDialog(true)}
+              />
             );
           })}
         </ul>
@@ -888,5 +318,7 @@ const InteractiveLegend = forwardRef<
     );
   }
 );
+
+InteractiveLegend.displayName = "InteractiveLegend";
 
 export default InteractiveLegend;
