--- app/src/components/Scatterplot/Scatterplot.module.css
@@ -1,64 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 1rem;
-  padding: 1rem;
-  background: var(--background);
-  border-radius: 0.5rem;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-}
-
-.controlsContainer {
-  display: flex;
-  gap: 1rem;
-  width: 100%;
-  padding: 0.5rem;
-  background: var(--background);
-  border-radius: 0.375rem;
-  border: 1px solid var(--border);
-}
-
-.errorMessage {
-  color: var(--error);
-  min-height: 1.5em;
-  padding: 0.5rem;
-  border-radius: 0.375rem;
-  background: rgba(239, 68, 68, 0.1);
-  width: 100%;
-  text-align: center;
-}
-
-.select {
-  padding: 0.5rem 0.75rem;
-  border: 1px solid var(--border);
-  border-radius: 0.375rem;
-  background-color: var(--background);
-  color: var(--foreground);
-  font-size: 0.875rem;
-  transition: all 0.2s ease-in-out;
-  cursor: pointer;
-}
-
-.select:hover {
-  border-color: var(--primary);
-}
-
-.select:focus {
-  outline: none;
-  border-color: var(--primary);
-  box-shadow: 0 0 0 2px var(--ring);
-}
-
-.scatterplot {
-  background-color: var(--background);
-  border: 1px solid var(--border);
-  border-radius: 0.5rem;
-  margin: 1rem;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-  transition: all 0.2s ease-in-out;
-}
-
-.scatterplot:hover {
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-}
