+++ app/src/components/StatusBar/StatusBar.types.ts
@@ -1,0 +1,8 @@
+export interface StatusBarProps {
+  totalProteins: number;
+  displayedProteins: number;
+  selectedProteins: number;
+  projectionName: string;
+}
+
+
